CS-355 Fixes: Attempt to delete image before deleting commands. Do not attempt to delete commands that are not saved.

+++ src/main/java/org/nrg/containers/services/impl/DockerServiceImpl.java
@@ -376,16 +376,22 @@ public class DockerServiceImpl implements DockerService {
 
     public void removeImageById(final String imageId, final Boolean force)
             throws NoDockerServerException, DockerServerException {
-        for (final DockerImageAndCommandSummary dockerImageAndCommandSummary : getImageSummaries()) {
+        final List<DockerImageAndCommandSummary> dockerImageAndCommandSummaries = getImageSummaries();
+
+        controlApi.deleteImageById(imageId, force);
+
+        for (final DockerImageAndCommandSummary dockerImageAndCommandSummary : dockerImageAndCommandSummaries) {
             if (dockerImageAndCommandSummary.imageId() != null &&
                     (dockerImageAndCommandSummary.imageId().equals(imageId) || dockerImageAndCommandSummary.imageId().contains(imageId))) {
                 for (final Command command : dockerImageAndCommandSummary.commands()) {
-                    commandService.delete(command.id());
+                    // The commands in the command summary list can have ID 0 if they are in the image labels but not saved in DB.
+                    if (command.id() != 0L) {
+                        commandService.delete(command.id());
+                    }
                 }
             }
         }
 
-        controlApi.deleteImageById(imageId, force);
     }
 
     @Override
